+++ src/Compilateur/AnalyseSyntaxique/Instruction/Fonction/ParseurRetour.cpp
@@ -2,7 +2,6 @@
 #include "Compilateur/AST/Noeuds/Interfaces/INoeud.h"
 #include "Compilateur/Lexer/TokenType.h"
 #include "Compilateur/AnalyseSyntaxique/Instruction/Fonction/ParseurRetour.h"
-#include <memory>
 #include <vector>
 
 ParseurRetour::ParseurRetour(IConstructeurArbre* constructeurEquation)
@@ -20,7 +19,7 @@ INoeud* ParseurRetour::parser(std::vector<Token>& tokens, int& index)
 
     INoeud* valeurRetour = nullptr;
 
-    if (index < (int)tokens.size() && tokens[index].type != TOKEN_POINT_VIRGULE) {
+    if (index < static_cast<int>(tokens.size()) && tokens[static_cast<size_t>(index)].type != TOKEN_POINT_VIRGULE) {
         valeurRetour = _constructeurEquation->construire(tokens, index);
     } else {
         consommer(tokens, index, TOKEN_POINT_VIRGULE, "Erreur: point-virgule attendu après return");
